+++ polars/polars-lazy/polars-plan/src/dsl/function_expr/schema.rs
@@ -292,6 +292,8 @@ impl FunctionExpr {
             }),
             #[cfg(feature = "log")]
             Entropy { .. } => float_dtype(),
+            #[cfg(feature = "log")]
+            Log1p => float_dtype(),
         }
     }
 }
